+++ src/components/Catalogs.vue
@@ -41,7 +41,7 @@
       <div :class="view === 'list' ? 'card-list' : 'card-grid'">
         <Catalog v-for="catalog in catalogView" :catalog="catalog" :viewMode="view" :key="catalog.href">
           <template #footer="{data}">
-            <slot name="catalogFooter" :data="data" />
+            <slot name="footer" :data="data" />
           </template>
         </Catalog>
       </div>