+++ src/components/Item.vue
@@ -15,6 +15,11 @@
       <Keywords v-if="showKeywordsInItemCards && keywords.length > 0" :keywords="keywords" variant="primary" />
       <b-card-text><small class="datetime" v-html="displayTime" /></b-card-text>
     </b-card-body>
+    <b-card-footer>
+      <slot name="footer" :data="data">
+        <StacActions v-if="data" :data="data" variant="outline-primary" vertical size="sm" />
+      </slot>
+    </b-card-footer>
   </b-card>
 </template>
 
@@ -25,17 +30,20 @@ import { mapState, mapGetters } from 'vuex';
 import FileFormatsMixin from './FileFormatsMixin';
 import CardMixin from './CardMixin';
 import StacLink from './StacLink.vue';
+import StacActions from './StacActions.vue';
 import { STAC } from 'stac-js';
 import { formatTemporalExtent, formatTimestamp } from '@radiantearth/stac-fields/formatters';
-import { BCard, BCardBody, BCardText, BCardTitle, BCardImg } from 'bootstrap-vue-next';
+import { BCard, BCardBody, BCardFooter, BCardText, BCardTitle, BCardImg } from 'bootstrap-vue-next';
 
 export default defineComponent({
   name: 'Item',
   components: {
     StacLink,
+    StacActions,
     BCard,
     BCardImg,
     BCardBody,
+    BCardFooter,
     BCardText,
     BCardTitle,
     Keywords: defineAsyncComponent(() => import('./Keywords.vue'))