+++ src/views/ApiSearch.vue
@@ -43,14 +43,15 @@
             :count="totalCount" :apiFilters="collectionFilters"
             @click="armCarryOver"
           >
-            <template #catalogFooter="slot">
-              <b-button-group v-if="itemSearch || canFilterItems(slot.data)" vertical size="sm">
+            <template #footer="slot">
+              <b-button-group v-if="itemSearch || canFilterItems(slot.data) || hasStacActions(slot.data)" vertical size="sm">
                 <b-button v-if="itemSearch" variant="outline-primary" :pressed="selectedCollections[slot.data.id]" @click="selectForItemSearch(slot.data)">
                   <b-icon-check-square v-if="selectedCollections[slot.data.id]" />
                   <b-icon-square v-else />
                   <span class="ms-2">{{ $t('search.selectForItemSearch') }}</span>
                 </b-button>
                 <StacLink :button="{variant: 'outline-primary', disabled: !canFilterItems(slot.data)}" :data="slot.data" :title="$t('search.filterCollection')" :state="{itemFilterOpen: 1}" />
+                <StacActions :data="slot.data" variant="outline-primary" size="sm" bare />
               </b-button-group>
             </template>
           </Catalogs>
@@ -84,6 +85,7 @@ import { STAC } from 'stac-js';
 import { defineComponent, defineAsyncComponent } from 'vue';
 import { getErrorCode, getErrorMessage, fetchQueryablesForLink, fetchSortablesForLink } from '../store/utils';
 import { mapGetters, mapState } from "vuex";
+import StacActionsConfig from '../../stacActions.config';
 import { BTab, BTabs } from 'bootstrap-vue-next';
 
 export default defineComponent({
@@ -97,6 +99,7 @@ export default defineComponent({
     SearchFilter,
     Items: defineAsyncComponent(() => import('../components/Items.vue')),
     MapView: defineAsyncComponent(() => import('../components/MapView.vue')),
+    StacActions: defineAsyncComponent(() => import('../components/StacActions.vue')),
     StacLink: defineAsyncComponent(() => import('../components/StacLink.vue'))
   },
   props: {
@@ -286,6 +289,9 @@ export default defineComponent({
         return Boolean(data.getApiItemsLink());
       }
       return false;
+    },
+    hasStacActions(data) {
+      return Object.values(StacActionsConfig).some(plugin => new plugin(data, this).show);
     },
     async loadResults(link) {
       this.error = null;